server/requirements.txt
Flask
flask-cors
pyscreenshot
pyautogui
pygetwindow
Pillow
requests
itsdangerous
Werkzeug
pywin32
